wrap: Do not return an fd from install_fd()
This fd is usually non-null, and breaks ioctl() through `if (ret)`
returning the fd instead of a possible failure code for the ioctl.

Since this way of "chaining" the fd through install_fd is not needed
anywhere, turn the only -1 branch into an assert and convert its return
type to void.